Update Angela.java to handle undo case
Add lastFunction local variable to Angela.java

Let's create a variable that stores the last inputted command that
was called by the user for use in the undo function.

This allows for the undo method to recall the last function used
and act accordingly in the future.

Add new undo() method to Angela.java

The method will recall the last function, calling an undo method in
the manager if the last function made changes to the database.

In the event of a bad undo (undoing a list, find command), the
method will give an error message.