Add undo functionality
Task class modifications

Task class and its subclasses are not effectively final.

This caused problems when archiving the previous iteration
of the task list.

Let's change do/undo task methods to return a new copy
of the task, allowing for easier archiving for the tasks.

Add support for archiving task list in TaskManager

Lets let TaskManager save a copy of the previous iteration
of tasks before any changes to the tasklist such that
it can be easily retrieved to undo the change.